Create a collector for each wakatime endpoint
