+++ collectors/summary/types.go
@@ -1,0 +1,135 @@
+/*
+Wakatime Exporter for Prometheus
+Copyright (C) 2020 Jacob Colvin (MacroPower)
+
+This program is free software; you can redistribute it and/or
+modify it under the terms of the GNU General Public License
+as published by the Free Software Foundation; either version 2
+of the License, or (at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program; if not, write to the Free Software
+Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+*/
+
+package summary
+
+import (
+	"io"
+	"net/url"
+	"sync"
+	"time"
+
+	"github.com/go-kit/kit/log"
+	"github.com/prometheus/client_golang/prometheus"
+)
+
+type exporterSummary struct {
+	URI       *url.URL
+	endpoint  string
+	user      string
+	mutex     sync.RWMutex
+	fetchStat func(url.URL, string, string) (io.ReadCloser, error)
+	tzOffset  time.Duration
+
+	up                          prometheus.Gauge
+	totalScrapes, queryFailures prometheus.Counter
+	logger                      log.Logger
+}
+
+type wakatimeSummary struct {
+	Data []struct {
+		Categories []struct {
+			Digital      string  `json:"digital"`
+			Hours        int     `json:"hours"`
+			Minutes      int     `json:"minutes"`
+			Name         string  `json:"name"`
+			Percent      float64 `json:"percent"`
+			Seconds      int     `json:"seconds"`
+			Text         string  `json:"text"`
+			TotalSeconds float64 `json:"total_seconds"`
+		} `json:"categories"`
+		Dependencies []struct {
+			Digital      string  `json:"digital"`
+			Hours        int     `json:"hours"`
+			Minutes      int     `json:"minutes"`
+			Name         string  `json:"name"`
+			Percent      float64 `json:"percent"`
+			Seconds      int     `json:"seconds"`
+			Text         string  `json:"text"`
+			TotalSeconds float64 `json:"total_seconds"`
+		} `json:"dependencies"`
+		Editors []struct {
+			Digital      string  `json:"digital"`
+			Hours        int     `json:"hours"`
+			Minutes      int     `json:"minutes"`
+			Name         string  `json:"name"`
+			Percent      float64 `json:"percent"`
+			Seconds      int     `json:"seconds"`
+			Text         string  `json:"text"`
+			TotalSeconds float64 `json:"total_seconds"`
+		} `json:"editors"`
+		GrandTotal struct {
+			Digital      string  `json:"digital"`
+			Hours        int     `json:"hours"`
+			Minutes      int     `json:"minutes"`
+			Text         string  `json:"text"`
+			TotalSeconds float64 `json:"total_seconds"`
+		} `json:"grand_total"`
+		Languages []struct {
+			Digital      string  `json:"digital"`
+			Hours        int     `json:"hours"`
+			Minutes      int     `json:"minutes"`
+			Name         string  `json:"name"`
+			Percent      float64 `json:"percent"`
+			Seconds      int     `json:"seconds"`
+			Text         string  `json:"text"`
+			TotalSeconds float64 `json:"total_seconds"`
+		} `json:"languages"`
+		Machines []struct {
+			Digital       string  `json:"digital"`
+			Hours         int     `json:"hours"`
+			MachineNameID string  `json:"machine_name_id"`
+			Minutes       int     `json:"minutes"`
+			Name          string  `json:"name"`
+			Percent       float64 `json:"percent"`
+			Seconds       int     `json:"seconds"`
+			Text          string  `json:"text"`
+			TotalSeconds  float64 `json:"total_seconds"`
+		} `json:"machines"`
+		OperatingSystems []struct {
+			Digital      string  `json:"digital"`
+			Hours        int     `json:"hours"`
+			Minutes      int     `json:"minutes"`
+			Name         string  `json:"name"`
+			Percent      float64 `json:"percent"`
+			Seconds      int     `json:"seconds"`
+			Text         string  `json:"text"`
+			TotalSeconds float64 `json:"total_seconds"`
+		} `json:"operating_systems"`
+		Projects []struct {
+			Digital      string  `json:"digital"`
+			Hours        int     `json:"hours"`
+			Minutes      int     `json:"minutes"`
+			Name         string  `json:"name"`
+			Percent      float64 `json:"percent"`
+			Seconds      int     `json:"seconds"`
+			Text         string  `json:"text"`
+			TotalSeconds float64 `json:"total_seconds"`
+		} `json:"projects"`
+		Range struct {
+			Date     string    `json:"date"`
+			End      time.Time `json:"end"`
+			Start    time.Time `json:"start"`
+			Text     string    `json:"text"`
+			Timezone string    `json:"timezone"`
+		} `json:"range"`
+	} `json:"data"`
+	End   time.Time `json:"end"`
+	Start time.Time `json:"start"`
+}
